# server/openapi.yaml
openapi: '3.0.3'
info:
  title: ESA Dono API
  description: |
    Donation platform with Stripe checkout, wallet balance, pledge (smart cart)
    auto-fulfillment, rewards, polls, fund goals, channels, and moderator/admin tooling.

    ## Authentication

    Donor auth is held in an **httpOnly `dono_session` cookie** (ADR 0004) so the
    credential is never readable by JS. The browser sends it automatically.
    Non-browser/API clients may instead present the donor magic token as
    `Authorization: Bearer <magic_token>`. Operational keys (admin/moderator) are
    operator secrets carried as prefixed Bearer credentials. The legacy `?token=`
    query param and `X-Admin-Key`/`X-Moderator-Key` headers have been removed.

    | Method | Transport | Used By |
    |--------|-----------|---------|
    | Donor session (browser) | `dono_session` httpOnly cookie (SameSite=Lax) | Donor wallet, spend endpoints, moderator |
    | Donor token (API) | `Authorization: Bearer <magic_token>` (or `Bearer donor_<magic_token>`) | Scripts / non-browser donor auth |
    | Admin key | `Authorization: Bearer key_admin_<ADMIN_API_KEY>` | Admin routes; also a fallback for moderator routes |
    | Moderator key | `Authorization: Bearer key_mod_<MODERATOR_API_KEY>` | Fallback access to moderator routes, independent of donor roles |
    | Moderator/Admin role | Donor session/token + `role: MODERATOR`/`ADMIN` on donor record | Moderator routes (ADMIN implies MODERATOR access) |
...

...

    ## Rate Limiting

...

...

...
servers:
...
paths:
...
  /rewards/{id}/claim:
    post:
      tags: [Rewards]
      summary: Claim a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                claim_data: {}
                quantity:
                  type: integer
                  minimum: 1
                  default: 1
                  description: Number of units of this reward to claim in one call. Decrements quantity_total by this amount and creates one RewardClaim row per unit (claim_data is shared across all of them).
      responses:
        '200':
          description: Claim created
        '400':
          description: Insufficient balance / sold out / inactive
        '401':
          description: Token required
        '429':
          description: Rate limited

  # ─── Polls ───
...
  /auctions:
    get:
      tags: [Auctions]
      summary: List active auctions
      parameters:
        - name: channel_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Active auctions ordered by ends_at ascending
  /auctions/{id}:
    get:
      tags: [Auctions]
      summary: Get auction detail with recent bid history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Auction detail
        '404':
          description: Auction not found
  /auctions/{id}/bid:
    post:
      tags: [Auctions]
      summary: Place a bid on an open auction
      description: |
        Requires a verified email and at least one prior completed donation.
        No balance is deducted — bids are commitments settled by the
        post-close payment cascade.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_cents]
              properties:
                amount_cents: { type: integer }
      responses:
        '200':
          description: Bid placed

  # ─── Channels ───
...
  /admin/rewards:
    get:
      tags: [Admin]
      summary: List all rewards
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: All rewards ordered by created_at descending
    post:
      tags: [Admin]
      summary: Create a reward
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, cost_cents]
              properties:
                title: { type: string }
                description: { type: string }
                type:
                  type: string
                  enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM]
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean, default: true }
                custom_type_label: { type: string }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward created
  /admin/rewards/{id}:
    put:
      tags: [Admin]
      summary: Update a reward
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean }
                custom_type_label: { type: string }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward updated
    delete:
      tags: [Admin]
      summary: Delete a reward
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
...
  /admin/auctions:
    get:
      tags: [Admin]
      summary: List all auctions
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: All auctions ordered by created_at descending
    post:
      tags: [Admin]
      summary: Create an auction
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, starting_price_cents, min_increment_cents, ends_at]
              properties:
                title: { type: string }
                description: { type: string }
                type:
                  type: string
                  enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM]
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                starting_price_cents: { type: integer }
                min_increment_cents: { type: integer }
                ends_at: { type: string, format: date-time }
                is_active: { type: boolean, default: true }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Auction created
  /admin/auctions/{id}:
    put:
      tags: [Admin]
      summary: Update an auction
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                starting_price_cents: { type: integer }
                min_increment_cents: { type: integer }
                ends_at: { type: string, format: date-time }
                is_active: { type: boolean }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Auction updated
    delete:
      tags: [Admin]
      summary: Delete an auction (only if it has no bids)
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
        '409':
          description: Cannot delete an auction with existing bids
...
  /moderator/rewards:
    get:
      tags: [Moderator]
      summary: List all rewards
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Rewards ordered by created_at descending
    post:
      tags: [Moderator]
      summary: Create a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, cost_cents]
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean, default: true }
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward created
  /moderator/rewards/{id}:
    put:
      tags: [Moderator]
      summary: Update a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean }
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward updated
    delete:
      tags: [Moderator]
      summary: Delete a reward
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
...
  /moderator/auctions:
    get:
      tags: [Moderator]
      summary: List all auctions
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: All auctions ordered by created_at descending
    post:
      tags: [Moderator]
      summary: Create an auction
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, starting_price_cents, min_increment_cents, ends_at]
              properties:
                title: { type: string }
                description: { type: string }
                type:
                  type: string
                  enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM]
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                starting_price_cents: { type: integer }
                min_increment_cents: { type: integer }
                ends_at: { type: string, format: date-time }
                is_active: { type: boolean, default: true }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Auction created
  /moderator/auctions/{id}:
    put:
      tags: [Moderator]
      summary: Update an auction
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                custom_type_label: { type: string }
                image_url: { type: string, nullable: true }
                starting_price_cents: { type: integer }
                min_increment_cents: { type: integer }
                ends_at: { type: string, format: date-time }
                is_active: { type: boolean }
                channel_id: { type: string, nullable: true }
      responses:
        '200':
          description: Auction updated
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key matching `ADMIN_API_KEY` env var
...